Dear Eric,
Please find attached the patch of the changes we suggest to solve the bug related to Date time conversion when there are more than 3 digits in the fractional part of the seconds. The current fix checks if the fractional part had more than 3 digits and applies the fix. The existing code does not do this check if pattern is provided which could result in wrong time stamps being generated.


Thanks & Regards,
Guangming, Vadim & Hilda 


git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/Log/Trunk@34783 a68f48d4-8d00-424f-b812-3f22a534a02a